Moved api.js inside Assets/JS
